Move more implementations from headers to cpp.
