+++ src/Access/Common/AllowedClientHosts.cpp
@@ -5,9 +5,15 @@
 #include <Functions/likePatternToRegexp.h>
 #include <Poco/Net/SocketAddress.h>
 #include <Poco/RegularExpression.h>
+#include <boost/algorithm/string/predicate.hpp>
 #include <boost/algorithm/string/replace.hpp>
-#include <ifaddrs.h>
+#include <boost/range/algorithm/find.hpp>
+#include <boost/range/algorithm_ext/erase.hpp>
 #include <Common/DNSResolver.h>
+#include <ifaddrs.h>
+#include <filesystem>
+
+namespace fs = std::filesystem;
 
 
 namespace DB
@@ -183,6 +189,262 @@ namespace
 
         name = pattern;
     }
+}
+
+
+void AllowedClientHosts::IPSubnet::set(const IPAddress & prefix_, const IPAddress & mask_)
+{
+    prefix = prefix_;
+    mask = mask_;
+
+    if (prefix.family() != mask.family())
+    {
+        if (prefix.family() == IPAddress::IPv4)
+            prefix = IPAddress("::ffff:" + prefix.toString());
+
+        if (mask.family() == IPAddress::IPv4)
+            mask = IPAddress(96, IPAddress::IPv6) | IPAddress("::ffff:" + mask.toString());
+    }
+
+    prefix = prefix & mask;
+
+    if (prefix.family() == IPAddress::IPv4)
+    {
+        if ((prefix & IPAddress{8, IPAddress::IPv4}) == IPAddress{"127.0.0.0"})
+        {
+            // 127.XX.XX.XX -> 127.0.0.1
+            prefix = IPAddress{"127.0.0.1"};
+            mask = IPAddress{32, IPAddress::IPv4};
+        }
+    }
+    else
+    {
+        if ((prefix & IPAddress{104, IPAddress::IPv6}) == IPAddress{"::ffff:127.0.0.0"})
+        {
+            // ::ffff:127.XX.XX.XX -> ::1
+            prefix = IPAddress{"::1"};
+            mask = IPAddress{128, IPAddress::IPv6};
+        }
+    }
+}
+
+void AllowedClientHosts::IPSubnet::set(const IPAddress & prefix_, size_t num_prefix_bits)
+{
+    set(prefix_, IPAddress(num_prefix_bits, prefix_.family()));
+}
+
+void AllowedClientHosts::IPSubnet::set(const IPAddress & address)
+{
+    set(address, address.length() * 8);
+}
+
+AllowedClientHosts::IPSubnet::IPSubnet(const String & str)
+{
+    size_t slash = str.find('/');
+    if (slash == String::npos)
+    {
+        set(IPAddress(str));
+        return;
+    }
+
+    IPAddress new_prefix{String{str, 0, slash}};
+    String mask_str(str, slash + 1, str.length() - slash - 1);
+    bool only_digits = (mask_str.find_first_not_of("0123456789") == std::string::npos);
+    if (only_digits)
+        set(new_prefix, std::stoul(mask_str));
+    else
+        set(new_prefix, IPAddress{mask_str});
+}
+
+String AllowedClientHosts::IPSubnet::toString() const
+{
+    unsigned int prefix_length = mask.prefixLength();
+    if (isMaskAllBitsOne())
+        return prefix.toString();
+    else if (IPAddress{prefix_length, mask.family()} == mask)
+        return fs::path(prefix.toString()) / std::to_string(prefix_length);
+    else
+        return fs::path(prefix.toString()) / mask.toString();
+}
+
+bool AllowedClientHosts::IPSubnet::isMaskAllBitsOne() const
+{
+    return mask == IPAddress(mask.length() * 8, mask.family());
+}
+
+
+void AllowedClientHosts::clear()
+{
+    addresses = {};
+    subnets = {};
+    names = {};
+    name_regexps = {};
+    like_patterns = {};
+    any_host = false;
+    local_host = false;
+}
+
+bool AllowedClientHosts::empty() const
+{
+    return !any_host && !local_host && addresses.empty() && subnets.empty() && names.empty() && name_regexps.empty() && like_patterns.empty();
+}
+
+void AllowedClientHosts::addAddress(const IPAddress & address)
+{
+    if (address.isLoopback())
+        local_host = true;
+    else if (boost::range::find(addresses, address) == addresses.end())
+        addresses.push_back(address);
+}
+
+void AllowedClientHosts::removeAddress(const IPAddress & address)
+{
+    if (address.isLoopback())
+        local_host = false;
+    else
+        boost::range::remove_erase(addresses, address);
+}
+
+void AllowedClientHosts::addSubnet(const IPSubnet & subnet)
+{
+    if (subnet.getMask().isWildcard())
+        any_host = true;
+    else if (subnet.isMaskAllBitsOne())
+        addAddress(subnet.getPrefix());
+    else if (boost::range::find(subnets, subnet) == subnets.end())
+        subnets.push_back(subnet);
+}
+
+void AllowedClientHosts::removeSubnet(const IPSubnet & subnet)
+{
+    if (subnet.getMask().isWildcard())
+        any_host = false;
+    else if (subnet.isMaskAllBitsOne())
+        removeAddress(subnet.getPrefix());
+    else
+        boost::range::remove_erase(subnets, subnet);
+}
+
+void AllowedClientHosts::addName(const String & name)
+{
+    if (boost::iequals(name, "localhost"))
+        local_host = true;
+    else if (boost::range::find(names, name) == names.end())
+        names.push_back(name);
+}
+
+void AllowedClientHosts::removeName(const String & name)
+{
+    if (boost::iequals(name, "localhost"))
+        local_host = false;
+    else
+        boost::range::remove_erase(names, name);
+}
+
+void AllowedClientHosts::addNameRegexp(const String & name_regexp)
+{
+    if (boost::iequals(name_regexp, "localhost"))
+        local_host = true;
+    else if (name_regexp == ".*")
+        any_host = true;
+    else if (boost::range::find(name_regexps, name_regexp) == name_regexps.end())
+        name_regexps.push_back(name_regexp);
+}
+
+void AllowedClientHosts::removeNameRegexp(const String & name_regexp)
+{
+    if (boost::iequals(name_regexp, "localhost"))
+        local_host = false;
+    else if (name_regexp == ".*")
+        any_host = false;
+    else
+        boost::range::remove_erase(name_regexps, name_regexp);
+}
+
+void AllowedClientHosts::addLikePattern(const String & pattern)
+{
+    if (boost::iequals(pattern, "localhost") || (pattern == "127.0.0.1") || (pattern == "::1"))
+        local_host = true;
+    else if ((pattern == "%") || (pattern == "0.0.0.0/0") || (pattern == "::/0"))
+        any_host = true;
+    else if (boost::range::find(like_patterns, pattern) == name_regexps.end())
+        like_patterns.push_back(pattern);
+}
+
+void AllowedClientHosts::removeLikePattern(const String & pattern)
+{
+    if (boost::iequals(pattern, "localhost") || (pattern == "127.0.0.1") || (pattern == "::1"))
+        local_host = false;
+    else if ((pattern == "%") || (pattern == "0.0.0.0/0") || (pattern == "::/0"))
+        any_host = false;
+    else
+        boost::range::remove_erase(like_patterns, pattern);
+}
+
+void AllowedClientHosts::addLocalHost()
+{
+    local_host = true;
+}
+
+void AllowedClientHosts::removeLocalHost()
+{
+    local_host = false;
+}
+
+void AllowedClientHosts::addAnyHost()
+{
+    clear();
+    any_host = true;
+}
+
+void AllowedClientHosts::add(const AllowedClientHosts & other)
+{
+    if (other.containsAnyHost())
+    {
+        addAnyHost();
+        return;
+    }
+    if (other.containsLocalHost())
+        addLocalHost();
+    for (const IPAddress & address : other.getAddresses())
+        addAddress(address);
+    for (const IPSubnet & subnet : other.getSubnets())
+        addSubnet(subnet);
+    for (const String & name : other.getNames())
+        addName(name);
+    for (const String & name_regexp : other.getNameRegexps())
+        addNameRegexp(name_regexp);
+    for (const String & like_pattern : other.getLikePatterns())
+        addLikePattern(like_pattern);
+}
+
+void AllowedClientHosts::remove(const AllowedClientHosts & other)
+{
+    if (other.containsAnyHost())
+    {
+        clear();
+        return;
+    }
+    if (other.containsLocalHost())
+        removeLocalHost();
+    for (const IPAddress & address : other.getAddresses())
+        removeAddress(address);
+    for (const IPSubnet & subnet : other.getSubnets())
+        removeSubnet(subnet);
+    for (const String & name : other.getNames())
+        removeName(name);
+    for (const String & name_regexp : other.getNameRegexps())
+        removeNameRegexp(name_regexp);
+    for (const String & like_pattern : other.getLikePatterns())
+        removeLikePattern(like_pattern);
+}
+
+
+bool operator ==(const AllowedClientHosts & lhs, const AllowedClientHosts & rhs)
+{
+    return (lhs.any_host == rhs.any_host) && (lhs.local_host == rhs.local_host) && (lhs.addresses == rhs.addresses)
+        && (lhs.subnets == rhs.subnets) && (lhs.names == rhs.names) && (lhs.name_regexps == rhs.name_regexps)
+        && (lhs.like_patterns == rhs.like_patterns);
 }
