+++ src/Interpreters/Access/InterpreterGrantQuery.cpp
@@ -11,6 +11,7 @@
 #include <Interpreters/executeDDLQueryOnCluster.h>
 #include <boost/range/algorithm/copy.hpp>
 #include <boost/range/algorithm/set_algorithm.hpp>
+#include <boost/range/algorithm_ext/erase.hpp>
 
 namespace DB
 {
